style(architecture): Use Lightweight Architecture Decision Records
Adapting ADR as described by ThoughtWorkds Technology Radar: Much documentation can be replaced with
highly readable code and tests. In a world of evolutionary architecture, however, it's important to
record certain design decisions for the benefit of future team members as well as for external
oversight. Lightweight Architecture Decision Records is a technique for capturing important
architectural decisions along with their context and consequences. We recommend storing these
details in source control, instead of a wiki or website, as then they can provide a record that
remains in sync with the code itself. For most projects, we see no reason why you wouldn't want to
use this technique.

fix #8